refactor(app): extract window types and find_window_id into app/window.rs
Pulls WindowKind, WindowState, EntityWindowState, and find_window_id
out of snapdash.rs into a sibling app/window.rs. find_window_id was an
associated function on Snapdash but never touched &self — converting it
to a free function drops the Snapdash:: call-site noise.

mod.rs re-exports each name explicitly rather than glob (`pub use *`)
to avoid ambiguous re-exports between snapdash and window.
