+++ src/app/mod.rs
@@ -1,3 +1,5 @@
 mod snapdash;
+mod window;
 
-pub use snapdash::*;
+pub use snapdash::{FocusDirection, Message, SettingsSensor, Snapdash};
+pub use window::{EntityWindowState, WindowKind, WindowState, find_window_id};
